+++ src/serena/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2025-06-21"
+__version__ = "0.1.3"
 
 import logging