+++ src/solidlsp/ls_config.py
@@ -3,8 +3,10 @@ Configuration objects for language servers
 """
 
 import fnmatch
+from collections.abc import Iterable
 from dataclasses import dataclass, field
 from enum import Enum
+from typing import Self
 
 
 class FilenameMatcher:
@@ -37,18 +39,42 @@ class Language(str, Enum):
     DART = "dart"
     CPP = "cpp"
     PHP = "php"
+    CLOJURE = "clojure"
+    ELIXIR = "elixir"
+    TERRAFORM = "terraform"
+    # Experimental or deprecated Language Servers
     SWIFT = "swift"
+    TYPESCRIPT_VTS = "typescript_vts"
+    """Use the typescript language server through the natively bundled vscode extension via https://github.com/yioneko/vtsls"""
+    PYTHON_JEDI = "python_jedi"
+    """Jedi language server for Python (instead of pyright, which is the default)"""
+    CSHARP_OMNISHARP = "csharp_omnisharp"
+    """OmniSharp language server for C# (instead of the default csharp-ls by microsoft).
+    Currently has problems with finding references, and generally seems less stable and performant.
+    """
+
+    @classmethod
+    def iter_all(cls, include_experimental: bool = False) -> Iterable[Self]:
+        for lang in cls:
+            if include_experimental or not lang.is_experimental():
+                yield lang
+
+    def is_experimental(self) -> bool:
+        """
+        Check if the language server is experimental or deprecated.
+        """
+        return self in {self.TYPESCRIPT_VTS, self.PYTHON_JEDI, self.CSHARP_OMNISHARP}
 
     def __str__(self) -> str:
         return self.value
 
     def get_source_fn_matcher(self) -> FilenameMatcher:
         match self:
-            case self.PYTHON:
+            case self.PYTHON | self.PYTHON_JEDI:
                 return FilenameMatcher("*.py", "*.pyi")
             case self.JAVA:
                 return FilenameMatcher("*.java")
-            case self.TYPESCRIPT:
+            case self.TYPESCRIPT | self.TYPESCRIPT_VTS:
                 # see https://github.com/oraios/serena/issues/204
                 path_patterns = []
                 for prefix in ["c", "m", ""]:
@@ -56,7 +82,7 @@ class Language(str, Enum):
                         for base_pattern in ["ts", "js"]:
                             path_patterns.append(f"*.{prefix}{base_pattern}{postfix}")
                 return FilenameMatcher(*path_patterns)
-            case self.CSHARP:
+            case self.CSHARP | self.CSHARP_OMNISHARP:
                 return FilenameMatcher("*.cs")
             case self.RUST:
                 return FilenameMatcher("*.rs")
@@ -74,6 +100,12 @@ class Language(str, Enum):
                 return FilenameMatcher("*.php")
             case self.SWIFT:
                 return FilenameMatcher("*.swift")
+            case self.CLOJURE:
+                return FilenameMatcher("*.clj", "*.cljs", "*.cljc", "*.edn")  # codespell:ignore edn
+            case self.ELIXIR:
+                return FilenameMatcher("*.ex", "*.exs")
+            case self.TERRAFORM:
+                return FilenameMatcher("*.tf", "*.tfvars", "*.tfstate")
             case _:
                 raise ValueError(f"Unhandled language: {self}")